include detectors module
