+++ lib/rack/simple_user_agent.rb
@@ -1,5 +1,5 @@
 require "rack/simple_user_agent/version"
-require "rack/simple_user_agent/detector"
+require "rack/simple_user_agent/detectors"
 require 'rack/simple_user_agent/railtie' if defined?(Rails)
 
 module Rack
@@ -14,6 +14,6 @@ module Rack
   end
 
   class Request
-    include Rack::SimpleUserAgent::Detector
+    include Rack::SimpleUserAgent::Detectors
   end
 end
